Analysis Updates (primarily)
Reconfigured the default switchport mode analysis and added a dynamic auto/desirable section.  Changed the logic if there is no config as before it assumed it would automatically be in access.

Added a check for the number of local user accounts and give a warning if there's only one.

Reformatted the order of the general output section and used tabs to try and better align it.